backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.41
Flask-Cors==4.0.0
Werkzeug==2.2.3
PyMySQL==1.0.2
python-dotenv==0.19.0
PyJWT==2.1.0
cryptography==3.4.7
Pillow==10.4.0
flask-sock==0.6.0
simple-websocket==0.10.1
paramiko==3.4.0
Flask-Session==0.8.0
requests==2.31.0
alibabacloud-ecs20140526>=7.0.0
alibabacloud-domain20180129>=4.0.0
alibabacloud-cdn20180510>=3.0.0
PyYAML==6.0.1
